[FIX] update_module_names: Rename non updated XML-ID occurences
Leaving untouched the non update ir.model.data entries, there are
problems loading the database when the migration ends, as some
of the related records can be assets, views...

Renaming the entries to an unique name (generated from ID), we
assure Odoo's update process removed them when updating the target
module. If we would remove it them instead, we will need to remove
manually the related resources as well, so this is the best
strategy.