[ADD] rename_fields: New method for renaming completely fields
This, in contrast of ``rename_columns``, performs all the steps for
completely rename a field from one name to another. This is needed for
making a complete renaming of a field with all their side features:
translations, filters, exports...

Call this method whenever you are not performing a pure SQL column renaming
for other purposes (preserve a value for example).

This method performs also the SQL column renaming, so only one call is
needed.