[ADD] disable_invalid_filters
Avoid errors on migrated databases making a non-destructive disabling.

It analyzes all the existing active filters to check if they are still
correct. If not, they are disabled for avoiding errors when clicking on
them, or worse, if they are default filters when opening the model/action.

To be run at the base end-migration script for having a general scope. Only
assured to work on >= v8.