Fix logging with non-ascii characters in exception.
str() should never be used in Python 2 if you can not know the encoding
of the string.

In this case it can log and raise a UnicodeEncodeError in migration.log
instead of the exception causing the error, and you cannot know why
your migration failed.

I use repr() for Python 2. It's not perfect but the exception is
displayed and the stacktrace is in the log because of logger.exception(),
which is bypassed if str() cause enother exception.