Fix conversion to attachment on large datasets
Selecting all binary data at once will exceed memory limits, and get the
Odoo process killed.
Fetching one by one is not enough, since executing the SELECT to
retrieve all the table data can also fail with a PostgreSQL "out of
memory" error.
The solution is to process the files to convert in smaller batches of
500 records.