[FIX] merge_records: Handle better translatable fields
When merging records from a model that has translatable fields, before this
patch, the translations for all the records are reassigned to the target
record, provoking duplicated translations. This is even forgiven in v12.

With this new approach, target translation has priority, and if not present,
then the first found is assigned instead.

Fixes #134