[IMP] Add XML conversion framework and Bootstrap 3 to 4 converter
In `openupgrade_tools`, I added some methods to perform transformation of XML strings based on given specs. The specs support some HTML-specific features, such as classes and styles. I provide also a handy shortcut for spec declaration.

I also added `openupgrade_120` with the spec for the Bootstrap 3 to 4 migration and a simple method that leverages the framework and spec to convert an XML string. Main method is to keep a consistent API across openupgradelib method calls. Other private methods are simpler and implement the actual logic.