[FIX] Respect preexisting styles and attributes
Without this patch, if an XML node had a `style="height:333px"` and the rules set it to `style="height:222px"`, there was no way to avoid that.

Now, if you *really* want to replace a style or attribute, then just add it to the list of styles/attributes to remove. Otherwise, it will only be added if it was missing before.