Allow callable XML replacements
Some conversions need to know details about other styles to be able to convert them.

After this patch, all replacement arguments can be callables that will get the original node details as kwargs, allowing more customization.

The patch includes a required change that benefits from this improvement, to serve also as an example. Previous Odoo versions allowed the user to modify the height of the element, but in v12 it's done through the min-height style, so the translation needs to know if there was previously a stablished height.