[IMP] update_module_names: rename instead of merge case
Do a rename instead of merge if new_name doesn't exist yet. It may happens that you do a merge from an installed module to an uninstalled module, but the later is a new module (and thus is not yet in the ir.module.module list) and it is not installed.

Example:

In v11, you have customer_activity_statement and customer_outstanding_statement.
In v12, you have partner_statement, that both previous modules merge into.