[FIX] Adapt `logged_query` to work with `psycopg2.sql.SQL` and similar helpers
I found the [`psycopg2.sql`](http://initd.org/psycopg/docs/sql.html) module with lots of great features for a SQL-composing library such as openupgradelib, and I'm using it in the new method being added.

However, you can't just do `query % args` as it was being done before, because these are sql-specific objects, not raw strings, and don't support that. BTW, that was misleading because the logged query was not exactly the query that was being sent to the database. [As explained in psycopg2 docs](http://initd.org/psycopg/docs/usage.html?highlight=never#sql-injection):

> Warning
> Never, **never, NEVER** use Python string concatenation (+) or string parameters interpolation (%) to pass variables to a SQL query string. Not even at gunpoint.

A few lines above there was another logger that did it more accurately: `logger.error('Error running %s' % cr.mogrify(query, args))`. This returns the same string that is sent to Postgres server, so it's a better candidate for debugging. The problem with that one is that under Python 3 it's a `bytes()` object, which means that the log is not properly formatted (basically it prints a raw `\n` instead of a real newline character, making it very hard to read by humans), and that's why I'm using `.decode()` at the end, which works fine both in Python 2 and 3.

Odoo cursors store last executed query information in `cr._obj.query`, and since this method is called a lot, I'm using that attribute instead of calling `cr.mogrify()` if possible, but since it's a hidden and undocumented attribute, I added a fallback to use `cr.mogrify()` just in case old or newer Odoo versions lack the attribute.