[IMP] Speed up set_message_last_post
Before this patch, the chosen `table` was inspected row by row, with a subselect that was also executed row by row to find messages attached to that model.

With this patch, all messages' max dates related with the used table are searched just once, and then joined with the target `table` to find records to update.

The result is reducing complexity from exponential to plain.