[IMP] rename_fields: Tolerant to target field column existence
The target field SQL column may already exists due to several reasons:

- Remnant SQL column from testings.
- Existing field in previous Odoo versions.
- Others...

so it's advisable to not crash in these cases. We keep the column
renaming it, but prevent this crash and the expected renaming is
performed.

The only drawback of this is that if you are a migration script
developer, and you perform a fault calling to this method, you may not
be awared of the problem as it's silenced.

Fixes OCA/OpenUpgrade#2339