Add functions to autoupdate unmodified COW views
In Odoo v12+, with `website` installed, and with its COW system for view edits, when a user clicks enables or disables a view marked as `customize_show=True` by using the dropdown in the website editor, Odoo creates a COW website-specific view which is enabled or disabled as desired.

When updating, COW views are usually not updated. If they contain any logic, it can lead to template rendering problems easily.

These errors make sense on some circumstances, but if it was just because the user used the customize show feature, they are disturbing and meaningless. With these methods provided here, this problem can be reduced to only meaninful occurrences.

The 1s method will compare website-specific views with their agnostic counterparts. If bodies are equal, they will be marked.

The 2nd method takes all those marked views and resets their bodies to the new upstream, which possibly comes from the module and is updated for the new Odoo version.

@Tecnativa TT23136