[IMP] Defer invalidating filters
When bridging multiple versions during a migration, custom modules are
usually only available in the target version. If we run this method with
every version, filters that depend on fields from these modules will be
invalidated prematurely. Also, running the method to invalidate filters on
every version has a performance impact.

The target version can be provided with an environment variable

```
export OPENUPGRADE_TARGET_VERSION=13.0
```

Co-Authored-By: Pedro M. Baeza <pedro.baeza@tecnativa.com>