[IMP] _change_foreign_key_refs: Add extra_where params
Adding the new argument `extra_where`, we are able to change a FK
reference depending on an extra condition, so not all ocurrences are
replaced.

This is useful for example when we have a unique record for all
companies, and on new version, there should be one record per company.
We perform the split, and we need to replace references for only
a specific company (extra_where="AND company_id = X").

Retro-compatibility is kept as the argument is optional, and a safeguard
has been introduced for skipping models where the condition can't be
fulfilled if the field doesn't exist (so we are not able to say if
the reference should be replaced or not).

TT28115