[IMP] delete_records_safely_by_xml_id: Mark as noupdate failed ones
The original idea of this function was to force the removal of
noupdate=1 records when they disappear in a new version, as they are not
removed regularly by Odoo update machinery.

But there's a hole in the data removal when the record is noupdate=0,
as they are removed through the update process... if they can be
removed. Maybe such records has been used somewhere, and they can't be
removed due to restrict foreign keys, provoking a crash.

Using this method for trying to remove them on advance (before the
Odoo update cleanup process), and marking them as noupdate=1 if not
possible, provides a safest path for not having to deal with errors
when performing such regular update. As this is totally opt-in,
migration scripts developers can think if calling this method with
such kind of records provokes any side effect or not.

Example of usage in https://github.com/OCA/OpenUpgrade/commit/7773f07d1bca222e5d25fe71fb66e530f5e461d1