[FIX+IMP] _change_generic: Allow to transfer generic stuff between models + fix followers through SQL
Adding a new keyword argument allows us to pass a possible new model for
using this function in other contexts like transferring the generic
stuff from one model to another (e.g. from account.invoice to
account.move in the v12 > v13 migration).

When adding this feature, I have discovered as well a bug in current
code using the SQL method, where barely no followers where transferred,
as the query was not having into account the model, so as soon as the
partner is follower of any record of any model with such ID, then it's
not transferred.