[IMP] convert_html_fragment: prevent lxml wraps and false positives
When the fragment has no common node, lxml wraps the code under a common one.
This content for instance:

    `<p><p/><p></p>`

is parsed as:

`<div><p><p/><p></p></div>`

To avoid this we force a custom wrapper tag on every parsed string so every xml
receives the same treatment and we can extract it later with no harm.

We don't want to update any fragment which has no changes after all the
replacements are checked. For that cases we'll return the original string.

At the end, we just trim our initial custom wrapper tag and return our
treated string.

TT44169 TT45734