[FIX] migrate_translations_to_jsonb: avoid ir_translation terms cleanup
We're making use of core's `_get_translation_upgrade_queries` which
aside from provide the proper queries to upgrade the translated fields
to the new json structure tries to cleanup the former ir_tanslation
records.

In OpenUpgrade we try to keep historic info after migration so we're
able to quickly react to possible inconsistencies.

Besides that, we need to rename the `ir_translation` table if we want to
work with this method.

TT49615