[IMP] merge_records: delete before update
In some case we might be intereseted in delete the origin records before
updating the target with the gathered values to avoid some possible
constraints like unique key values.

For example, when updating product_product barcode with a field
specification of first_from_origin, the merge will fail as barcode is a
unique key for that table.

To change this behavior, a new key `openupgradelib_delete_before_update`
can be added to the field_spec dict.