[IMP] openupgrade_180.py: check table ir_property exists
If you are migrating using the enterprise service, the table ir_property is already converted and deleted. Also the conversion to company_dependent is done for all the fields in that service too. So when trying to access the table it raises an exception. So I'm adding a condition to ensure this table still exists and being able to do the migration with both services. For reference, check: https://github.com/OCA/purchase-workflow/pull/2861